requirements.txt
biopython==1.77
Flask==1.1.2
pandas==1.1.0